requirements.txt
# ============================================================================
# CORE DEPENDENCIES
# ============================================================================

# LLM & AI Libraries
langchain==0.1.0
google-generativeai>=0.3.0

# Vector Databases & Retrieval
faiss-cpu>=1.7.4
chromadb==0.4.24
sentence-transformers==2.2.2

# Data Processing & NLP
scikit-learn>=1.3.0
numpy==1.24.3
pandas>=2.0.0
nltk>=3.8.0
Pillow>=10.0.0

# Lexical Search
rank-bm25>=0.2.0

# ============================================================================
# WEB FRAMEWORKS & API
# ============================================================================

# Flask API Server
flask>=3.0.0
flask-cors>=4.0.0

# Streamlit Frontend
streamlit>=1.28.0

# ============================================================================
# UTILITIES & CONFIGURATION
# ============================================================================

# Environment & Config
python-dotenv>=1.0.0
pydantic>=2.0.0
typing-extensions>=4.0.0

# Data Processing
python-dateutil>=2.8.0
tqdm>=4.66.0
requests>=2.31.0

# ============================================================================
# DATABASE (OPTIONAL)
# ============================================================================

# pymongo>=4.5.0
# psycopg2-binary>=2.9.0

# ============================================================================
# TESTING & QUALITY
# ============================================================================

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0

# Code Quality
black>=23.0.0
flake8>=6.0.0
pylint>=3.0.0

# ============================================================================
# LOGGING
# ============================================================================

python-json-logger>=2.0.0
